+++ app/Models/PurchaseItem.php
@@ -33,7 +33,7 @@ class PurchaseItem extends Model
     public function price(): Attribute
     {
         return new Attribute(
-            get: fn ($value) => (float) to_rands($value),
+            get: fn ($value) => to_rands($value),
             set: fn ($value) => (float) to_cents($value)
         );
     }
