+++ app/Models/StockTakeItem.php
@@ -13,7 +13,7 @@ class StockTakeItem extends Model
     public function cost(): Attribute
     {
         return new Attribute(
-            get: fn ($value) => (float) to_rands($value),
+            get: fn ($value) => to_rands($value),
             set: fn ($value) => to_cents($value)
         );
     }
