+++ app/Models/SupplierCreditItem.php
@@ -27,7 +27,7 @@ class SupplierCreditItem extends Model
     public function cost(): Attribute
     {
         return new Attribute(
-            get: fn ($value) => (float) to_rands($value),
+            get: fn ($value) => to_rands($value),
             set: fn ($value) => to_cents($value)
         );
     }
